remove white spaces from Dashboard tables
